# alternator: yield while converting large responses to json text