We have two paths for generating the json text representation, one for large items and one for small items, but the large item path is lacking:

 - it doesn't yield, so a response with many items will stall
 - it doesn't wait for network sends to be accepted by the network stack, so it will allocate a lot of memory

Fix by moving the generation to a thread. This allows us to wait for the network stack, which incidentally also fixes stalls.

The cost of the thread is amortized by the fact we're emitting a large response.

Fixes #18806

No backport required: not a regression.