@nyh 2024-05-21
This is very confusing. https://github.com/scylladb/scylladb/pull/12351 by @nuivall  was all about adding yields to rjson::print, and it was done by using coroutines instead of continuations, and adding co_await in the right places to allow them to yield. So this problem should have been solved already. How did it reappear? Was it really solved in the first place? :-(