@nuivall 2024-05-21
> Ideas on how to test the new stalllessness are welcome.

`def test_deeply_nested_put` ?