@nuivall 2024-05-21
> > Ideas on how to test the new stalllessness are welcome.
> 
> `def test_deeply_nested_put` ? or `def test_batch_write_item_large` I see back then I've used it

or `./build/release/scylla perf-alternator --workdir ~/tmp --smp 1 --developer-mode 1 --alternator-port 8000 --alternator-write-isolation only_rmw_uses_lwt --workload scan --duration 10 --scan-total-segments 10 2> /dev/null
`