@avikivity 2024-05-21
> > This is very confusing. #12351 by @nuivall was all about adding yields to rjson::print, and it was done by using coroutines instead of continuations, and adding co_await in the right places to allow them to yield. So this problem should have been solved already. How did it reappear? Was it really solved in the first place? :-(
> 
> It wasn't solved. The json visitor is called in a single continuation. The writes to the output_stream are chained together, but don't pause the generation process.

A visitor that can yield (by storing its state in a stack and resuming) would be better, but much more work.