@nuivall 2024-05-21
> Do these generate stalls now?

Haven't checked, likely you need to fiddle parameters to make responses bigger.

> > This is very confusing. #12351 by @nuivall was all about adding yields to rjson::print, and it was done by using coroutines instead of continuations, and adding co_await in the right places to allow them to yield. So this problem should have been solved already. How did it reappear? Was it really solved in the first place? :-(
> 
> It wasn't solved. The json visitor is called in a single continuation. The writes to the output_stream are chained together, but don't pause the generation process.

I think it was solved in a meaning that it pushed the size of request which doesn't stall, but due to this chaining as we can see it didn't eliminate it completely. And back then if I recall the assumption was that we don't want to have thread in the response path.