@nyh 2024-05-21
> > This is very confusing. #12351 by @nuivall was all about adding yields to rjson::print, and it was done by using coroutines instead of continuations, and adding co_await in the right places to allow them to yield. So this problem should have been solved already. How did it reappear? Was it really solved in the first place? :-(
> 
> It wasn't solved. The json visitor is called in a single continuation. The writes to the output_stream are chained together, but don't pause the generation process.

I see now what you mean :-(
We definitely need a test to reproduce this problem, it should be easy to do, I can do this tomorrow.

I am really not happy about the seastar::thread solution, but if I understand correctly, it will only get used for is_big() (over 100KB) responses, so hopefully at that size the extra seastar::thread will not be so bad. @nuivall can you use your benchmark with large requests to measure  the performance degradation that this patch causes?