@nuivall 2024-05-22
> > > This is very confusing. #12351 by @nuivall was all about adding yields to rjson::print, and it was done by using coroutines instead of continuations, and adding co_await in the right places to allow them to yield. So this problem should have been solved already. How did it reappear? Was it really solved in the first place? :-(
> > 
> > 
> > It wasn't solved. The json visitor is called in a single continuation. The writes to the output_stream are chained together, but don't pause the generation process.
> 
> I see now what you mean :-( We definitely need a test to reproduce this problem, it should be easy to do, I can do this tomorrow.
> 
> I am really not happy about the seastar::thread solution, but if I understand correctly, it will only get used for is_big() (over 100KB) responses, so hopefully at that size the extra seastar::thread will not be so bad. @nuivall can you use your benchmark with large requests to measure the performance degradation that this patch causes?

BEFORE
```
⤖  scylladb git:(a86fb293f) ./build/release/scylla perf-alternator --workdir ~/tmp --smp 1 --developer-mode 1 --alternator-port 8000 --alternator-write-isolation only_rmw_uses_lwt --workload scan --duration 10 --scan-total-segments 10 2> /dev/null
Running test with config: {workload=scan, partitions=10000, concurrency=100, duration_in_seconds=10, operations-per-shard=0, flush=1}
Creating 10000 partitions...
Flushing partitions...
198.86 tps (183643.0 allocs/op,   0.6 logallocs/op, 634.6 tasks/op, 94921294 insns/op,        0 errors)
198.68 tps (183823.4 allocs/op,   0.5 logallocs/op, 630.1 tasks/op, 95013714 insns/op,        0 errors)
200.10 tps (183371.8 allocs/op,   0.4 logallocs/op, 631.2 tasks/op, 94823611 insns/op,        0 errors)
204.56 tps (183159.5 allocs/op,   0.0 logallocs/op, 589.4 tasks/op, 94669276 insns/op,        0 errors)
204.07 tps (183691.8 allocs/op,   0.0 logallocs/op, 599.3 tasks/op, 94938957 insns/op,        0 errors)
203.20 tps (183559.3 allocs/op,   0.0 logallocs/op, 593.7 tasks/op, 94874487 insns/op,        0 errors)
203.08 tps (183870.2 allocs/op,   0.0 logallocs/op, 610.0 tasks/op, 95036484 insns/op,        0 errors)
204.34 tps (183205.3 allocs/op,   0.0 logallocs/op, 592.0 tasks/op, 94696217 insns/op,        0 errors)
203.59 tps (183745.6 allocs/op,   0.0 logallocs/op, 607.1 tasks/op, 94968905 insns/op,        0 errors)
203.51 tps (183681.4 allocs/op,   0.0 logallocs/op, 588.9 tasks/op, 94939543 insns/op,        0 errors)

median 203.51432249614635
median absolute deviation: 0.82
maximum: 204.56
minimum: 198.68

```
AFTER
```
⤖  scylladb git:(alternator-large-item-unstall-print) ./build/release/scylla perf-alternator --workdir ~/tmp --smp 1 --developer-mode 1 --alternator-port 8000 --alternator-write-isolation only_rmw_uses_lwt --workload scan --duration 10 --scan-total-segments 10 2> /dev/null
Running test with config: {workload=scan, partitions=10000, concurrency=100, duration_in_seconds=10, operations-per-shard=0, flush=1}
Creating 10000 partitions...
Flushing partitions...
189.74 tps (183220.1 allocs/op,   0.7 logallocs/op, 199.5 tasks/op, 100608735 insns/op,        0 errors)
190.98 tps (183482.7 allocs/op,   0.7 logallocs/op, 189.1 tasks/op, 100735449 insns/op,        0 errors)
194.95 tps (182853.7 allocs/op,   0.1 logallocs/op, 169.2 tasks/op, 100382801 insns/op,        0 errors)
196.90 tps (182701.9 allocs/op,   0.0 logallocs/op, 161.6 tasks/op, 100298416 insns/op,        0 errors)
196.11 tps (183155.3 allocs/op,   0.0 logallocs/op, 161.3 tasks/op, 100541621 insns/op,        0 errors)
196.25 tps (183142.8 allocs/op,   0.0 logallocs/op, 160.9 tasks/op, 100541009 insns/op,        0 errors)
196.26 tps (183501.7 allocs/op,   0.0 logallocs/op, 161.9 tasks/op, 100735945 insns/op,        0 errors)
196.87 tps (182973.1 allocs/op,   0.0 logallocs/op, 160.8 tasks/op, 100442058 insns/op,        0 errors)
196.47 tps (183008.0 allocs/op,   0.0 logallocs/op, 161.1 tasks/op, 100458737 insns/op,        0 errors)
196.35 tps (183377.1 allocs/op,   0.0 logallocs/op, 160.7 tasks/op, 100665301 insns/op,        0 errors)

median 196.25702844218566
median absolute deviation: 0.61
maximum: 196.90
minimum: 189.74
```

I don't see big difference. There is 6% insns/op bump though.